Move ClientEntity to entity module
